requirements.txt
streamlit
google-generativeai
python-dotenv
korean-lunar-calendar